Pull relevant book info from Google Books API request results.
The result set from the API call contains an array of books that
match the search, but the first item in the array seems to be
matching up correctly everytime so I'm going to simply extract
that item from the list, without doing any checking to see if it
actually matches the book I want. Down the line, may need to make
this more robust by sending along some other informatin in the url,
like the title of the book.