Display book image at books/:isbn13 route.
Still not confident in this method of displaying images, but it
is a starting point at least. One of the images doesn't work
when I query the Open Books API, so that's a problem. Maybe passing
the image link from the bestseller's list on the homepage to the
/books/:isbn13 route is the answer? That won't help for the more
general search functionality down the road, but at least it will
make consistency possible between the homepage and the bookpage.