Refactor .center class and add Custom CSS Properties for gutter.
The .center class was using 'margin: 0 auto', but that meant that
if I wanted to add a top (or bottom) margin to elements, I would
have to override this or create another class lower in the cascade
to do so. Instead, I changed it to 'margin-left, margin-right' which
is a better way to do this I think. Also started using CSS Custom
Properties, which allow you to declare a variable in the CSS and
reuse it in different selectors.